Move more items to utils and fix clasp open
Clasp open was trying to open an object, rather than the scriptID
Moved over manifestExists, saveProjectId, checkIfOnline, and getAPIFileType
